+++ scripts/our_mpeg/transform.py
@@ -3,7 +3,7 @@ from scipy.fftpack import dct, idct
 
 def transform(array):
     """
-    Args: 
+    Args:
         list(list(numpy.ndarray)): 2D list of uint8 numpy arrays
     Returns:
         list(list(numpy.ndarry)): the dct coefficients as float32 of each numpy array in the 2D list
@@ -22,7 +22,7 @@ def inverse_transform(array):
         list(list(numpy.ndarry)): the idct casted to uint8 of each numpy array in the 2D list
     """
     return [
-        [np.rint(idct(array[x][y], norm='ortho')).astype(np.uint8)
+        [np.rint(idct(array[x][y], norm='ortho')).astype(np.int16)
          for y in range(0, len(array[x]))]
          for x in range(0, len(array))
     ]
